fix type warning
 replace method that now is private

+++ Samples/Example_ImGui_Docking/main_Windows.cpp
@@ -108,9 +108,9 @@ BOOL InitInstance(HINSTANCE hInstance, int nCmdShow)
    hInst = hInstance; // Store instance handle in our global variable
 
    //PE: Center window at 7% smaller then screen.
-   float sizeadjust = 0.93;
-   float screen_width = GetSystemMetrics(SM_CXSCREEN);
-   float screen_height = GetSystemMetrics(SM_CYSCREEN);
+   float sizeadjust = 0.93f;
+   int screen_width = GetSystemMetrics(SM_CXSCREEN);
+   int screen_height = GetSystemMetrics(SM_CYSCREEN);
 	RECT rect = { (long)(screen_width -(screen_width * sizeadjust)), (long)(screen_height -(screen_height * sizeadjust)), (long)(screen_width * sizeadjust), (long)(screen_height * sizeadjust)};
    AdjustWindowRect(&rect, WS_OVERLAPPEDWINDOW, true);
